# openapi.yaml
openapi: 3.0.0
info:
  title: Vigilante
  description: |
    Software abierto para fiscalizar elecciones. Conteo de votos, anotaciones y registro de impugnaciones a nivel distrito, municipio, centro y mesa. Promoviendo transparencia y participación ciudadana en procesos electorales.
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html#license-text
  version: 1.0.0
paths:
  /health:
    get:
      summary: Display the health status of the project
      operationId: health
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Health'
  /login:
    post:
      summary: Retorna el token si las credenciales son validas
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Login'
        "403":
          description: Credenciales inválidas
  /departamentos:
    get:
      summary: Despliega los departamentos disponibles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Departamento'
  /departamentos/{departamento_uuid}:
    get:
      summary: Despliega el departamento especificado
      security:
        - bearerAuth: []
      parameters:
        - name: departamento_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Departamento'
        "404":
          description: No se encontró el departamento
  /departamentos/{departamento_uuid}/reset:
    put:
      summary: Reinicia el recuento de votos de un departamento y todos sus municipios, centros y mesas
      security:
        - bearerAuth: []
      parameters:
        - name: departamento_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Operación exitosa
        "404":
          description: No se encontró el departamento
  /departamentos/{departamento_uuid}/municipios:
    get:
      summary: Despliega los municipios de un departamento determinado
      security:
        - bearerAuth: []
      parameters:
        - name: departamento_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Municipio'
        "404":
          description: No se encontró el departamento
  /municipios/{municipio_uuid}:
    get:
      summary: Despliega el municipio especificado
      security:
        - bearerAuth: []
      parameters:
        - name: municipio_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Municipio'
        "404":
          description: No se encontró el municipio
  /municipios/{municipio_uuid}/centros:
    get:
      summary: Despliega los centros de un municipio determinado
      security:
        - bearerAuth: []
      parameters:
        - name: municipio_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Centro'
        "404":
          description: No se encontró el municipio
  /centros/{centro_uuid}:
    get:
      summary: Despliega el centro especificado
      security:
        - bearerAuth: []
      parameters:
        - name: centro_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Centro'
        "404":
          description: No se encontró el centro
  /centros/{centro_uuid}/mesas:
    get:
      summary: Despliega las mesas de un centro determinado
      security:
        - bearerAuth: []
      parameters:
        - name: centro_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mesa'
        "404":
          description: No se encontró el centro
  /mesas/{mesa_uuid}:
    get:
      summary: Despliega la mesa especificada
      security:
        - bearerAuth: []
      parameters:
        - name: mesa_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mesa'
        "404":
          description: No se encontró la mesa
  /mesas/{mesa_uuid}/recuentos:
    put:
      summary: Actualiza el recuento de votos de una mesa
      security:
        - bearerAuth: []
      parameters:
        - name: mesa_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                partidos:
                  type: array
                  items:
                    type: object
                    properties:
                      partido_uuid:
                        type: string
                        format: uuid
                      votos:
                        type: integer
      responses:
        "201":
          description: Operación exitosa
        "404":
          description: No se encontró la mesa
  /partidos:
    get:
      summary: Despliega los partidos disponibles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Partido'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Health:
      type: object
      properties:
        health:
          type: string
        projectName:
          type: string
        projectVersion:
          type: string
    Login:
      type: object
      properties:
        token:
          type: string
          format: JWT
    Departamento:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        nombre:
          type: string
        código:
          type: string
        recuentos:
          type: array
          items:
            $ref: '#/components/schemas/Recuento'
    Municipio:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        nombre:
          type: string
        código:
          type: string
        recuentos:
          type: array
          items:
            $ref: '#/components/schemas/Recuento'
    Centro:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        nombre:
          type: string
        código:
          type: string
        recuentos:
          type: array
          items:
            $ref: '#/components/schemas/Recuento'
    Mesa:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        número:
          type: string
        recuentos:
          type: array
          items:
            $ref: '#/components/schemas/Recuento'
    Recuento:
      type: object
      properties:
        partido_uuid:
          type: string
          format: uuid
        votos:
          type: integer
    Partido:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        nombre:
          type: string
        código:
          type: string
        color:
          type: string